Refactor OutNoteTable: Remove TableData components & rebuild as standalone component
