--- pkg/ui/src/InventoryTable/TableData/index.ts
@@ -1,3 +1,0 @@
-export { default as CoreBookTableData } from './CoreBookTableData.svelte';
-export { default as OptionalBookTableData } from './OptionalBookTableData.svelte';
-export { default as OutNoteTableData } from './OutNoteTableData.svelte';
