--- pkg/ui/src/InventoryTable/enums.ts
@@ -1,4 +1,0 @@
-export enum InventoryTableVariant {
-	Default = 'default',
-	OutNote = 'out note'
-}
