+++ pkg/ui/src/InventoryTable/index.ts
@@ -1,5 +1,4 @@
 export { default as InventoryTable } from './InventoryTable.svelte';
+export { default as OutNoteTable } from './OutNoteTable.svelte';
 
 export { createTable } from './table';
-
-export * from './enums';
